Refactor implementation #2789
